# dev/dataapi/openapi.yaml
openapi: 3.0.0
info:
  title: 'Swagger checkmate apis'
  description: "checkmate api.  You can find\nout more about api at http://subbieapp.com/"
  termsOfService: 'http://swagger.io/terms/'
  contact:
    email: dinhminhquoi@yahoo.com
  license:
    name: 'Apache 2.0'
    url: 'http://www.apache.org/licenses/LICENSE-2.0.html'
  version: 1.0.0
servers:
  -
    url: 'http://subbieapp.com/checkmate/api/'
    description: 'OpenApi host'
  -
    url: 'http://checkmate/'
    description: 'OpenApi host'
paths:
  /LoginController/registration:
    post:
      tags:
        - login
      summary: ''
      description: ''
      operationId: projectStep3
      parameters:
        -
          name: name
          in: query
          description: 'your username'
          required: true
          schema:
            type: string
        -
          name: email_id
          in: query
          description: 'your email_id'
          required: true
          schema:
            type: string
        -
          name: password
          in: query
          description: 'your password'
          required: true
          schema:
            type: string
        -
          name: location_id
          in: query
          description: 'your location_id'
          required: true
          schema:
            type: integer
            format: int32
        -
          name: device_token
          in: query
          description: 'your device_token'
          required: true
          schema:
            type: string
        -
          name: device_type
          in: query
          description: 'your device_type'
          required: true
          schema:
            type: integer
            format: int32
      responses:
        '200':
          description: 'successful operation'
        '400':
          description: ''
  /LoginController/login:
    post:
      tags:
        - login
      summary: ''
      description: ''
      operationId: login
      parameters:
        -
          name: email_id
          in: query
          description: 'your email_id'
          required: true
          schema:
            type: string
        -
          name: password
          in: query
          description: 'your password'
          required: true
          schema:
            type: string
        -
          name: device_token
          in: query
          description: 'your device_token'
          required: true
          schema:
            type: string
        -
          name: device_type
          in: query
          description: 'your device_type'
          required: true
          schema:
            type: integer
            format: int32
      responses:
        '200':
          description: 'successful operation'
        '400':
          description: ''
  /LoginController/logout:
    post:
      tags:
        - login
      summary: ''
      description: ''
      operationId: logout
      parameters:
        -
          name: access_token
          in: query
          description: ''
          required: true
          schema:
            type: string
        -
          name: user_id
          in: query
          description: ''
          schema:
            type: integer
            format: int32
      responses:
        '200':
          description: 'successful operation'
        '400':
          description: ''
...
  /ProjectController/projectStep3:
    post:
      tags:
        - project
      summary: ''
      description: ''
      operationId: projectStep3
      parameters:
        -
          name: access_token
          in: query
          required: true
          schema:
            type: string
        -
          name: user_id
          in: query
          description: ''
          required: true
          schema:
            type: integer
            format: int32
        -
          name: project_id
          in: query
          required: true
          schema:
            type: string
        -
          name: project_brief_scope
          in: query
          required: true
          schema:
            type: string
        -
          name: project_scope
          in: query
          required: true
          schema:
            type: string
        -
          name: project_site_map
          in: query
          required: true
          schema:
            type: string
        -
          name: project_picture
          in: query
          required: true
          schema:
            type: string
      responses:
        '200':
          description: 'successful operation'
        '400':
          description: ''
...